Refactor idea book API endpoints to improve error handling and validation. Update GET, PATCH, and DELETE methods to ensure proper user authorization and return relevant error messages. Simplify response structures in professional portal APIs for events, portfolio items, and projects. Enhance file upload endpoint with authentication checks and improved error responses.
