+++ apps/client/app/api/uploads/route.ts
@@ -1,7 +1,9 @@
-import { NextResponse } from 'next/server';
+import { NextRequest, NextResponse } from 'next/server';
+import { auth } from '@clerk/nextjs/server';
 import fs from 'fs';
 import path from 'path';
 import { randomUUID } from 'crypto';
+import { apiError } from '@/app/lib/resilient-api';
 
 /**
  * POST /api/uploads
@@ -11,9 +13,17 @@ import { randomUUID } from 'crypto';
  * NOTE:
  * - In production you may want to upload directly to cloud storage (S3/GCS) and return remote URLs.
  * - This implementation writes into 'public/uploads' so returned URLs are relative (e.g. /uploads/<file>).
+ * - Requires authentication via Clerk.
  */
 
-export async function POST(req: Request) {
+export async function POST(req: NextRequest) {
+  // Authentication check
+  const { userId } = await auth();
+  
+  if (!userId) {
+    return apiError('Unauthorized', 401);
+  }
+
   try {
     const form = await req.formData();
     const uploadDir = path.join(process.cwd(), 'public', 'uploads');
@@ -43,6 +53,6 @@ export async function POST(req: Request) {
     return NextResponse.json({ success: true, uploaded });
   } catch (err) {
     console.error('Upload error:', err);
-    return NextResponse.json({ success: false, error: String(err) }, { status: 500 });
+    return apiError('File upload failed', 500, String(err));
   }
 }
